KNNEigenface/requirements.txt
numpy==1.19.4
Pillow==8.0.0
